Add hermetic chat test helpers
